Create Step1 - create_monthly_databaste.R
This script creates the main database but at monthly level, this code is necessary if we want to do the analysis using monthly outcomes. We can modify this code to create or include different controls from the different datasets. The final database from this script, as well as in the quarterly case, corresponds to a balanced monthly panel from 2012-1 to 20219-12 where we observe all individuals in all the periods.